First integration of new add-on code
New add-on code incorporates new identity strategy where a user holds
two identities concurrently.  One that it is actively using for
requests to google and one that is held as a backup.  When the current
identity expires it checks it back in to the server for someone else to
use and promotes the backup identity as the new current.  Lastly, it
fetches a new backup identity.